Splitted Controller
