--- src/main/java/mops/termine2/controller/AbstimmungTerminController.java
@@ -1,12 +1,0 @@
-package mops.termine2.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.context.annotation.SessionScope;
-
-@Controller
-@SessionScope
-@RequestMapping("/termine2/termine_abstimmung")
-public class AbstimmungTerminController {
-
-}
